longarr phase 2+3+4: 64-bit arrays and tables — daslang surface + AOT helpers
Widens runtime Array / Table structs (size, capacity, tombstones) to uint64
and lands the matching daslang surface — long_length / long_capacity,
int64 overloads of resize / reserve / erase / push, arr[i64] / arr[u64]
indexing, long_iter_range / long_enumerate / long_find_index, and int64
overloads of array_view / temp_array.

Phase 2c AOT helpers are folded in because Phase 4b's arr[i64] dispatch
immediately emits AOT code that references the new int64 / uint64
TArray::operator() and das_ati_i64 / das_atu_u64 / das_atci_i64 /
das_atcu_u64 free-function helpers — shipping Phase 4b without these
would break every AOT compile.

Daslib mirror tier updated to track the widened C++ virtuals:
DapiArray.size / DapiArray.capacity widened to uint64, DapiTable.tombstones
added, DapiDataWalker callbacks (count / index params) widened to uint64.
daslib/debug.das and daslib/debug_eval.das override sigs updated to match.

length() / capacity() panic via DAS_VERIFYF when the underlying uint64 size
exceeds INT_MAX — the int-returning surface stays binary-compatible but
loud-fails on overflow rather than silently truncating.

LLVM JIT codegen version bumped 0x02 to 0x03 so stale JIT DLLs auto-GC
on next run.

Tests: 16 new under tests/long_array_table/ covering long_length,
arr[i64] / arr[u64] indexing, int64 overloads, long iterators —
registered in tests/aot/CMakeLists.txt for AOT coverage.

Lint sweep on tests/data_walker/: STYLE017 (`if cond return false; return true`
collapsed to `return !cond`) and 5 LINT003 (`var` to `let`) fixes.

Plan tracker: C:\Users\Boris\.claude\plans\swift-rolling-pelican.md.

+++ include/daScript/daScriptC.h
@@ -104,8 +104,8 @@ typedef struct {
 
 typedef struct {
     char *      data;
-    uint32_t    size;
-    uint32_t    capacity;
+    uint64_t    size;
+    uint64_t    capacity;
     uint32_t    magic;
     uint32_t    lock;
     uint32_t    flags;
@@ -114,14 +114,14 @@ typedef struct {
 // Layout MUST mirror das::Table exactly (Table : Array, then keys/hashes/tombstones).
 typedef struct {
     char *      data;
-    uint32_t    size;
-    uint32_t    capacity;
+    uint64_t    size;
+    uint64_t    capacity;
     uint32_t    magic;
     uint32_t    lock;
     uint32_t    flags;
     char *      keys;
     uint32_t *  hashes;
-    uint32_t    tombstones;
+    uint64_t    tombstones;
 } das_table;
 
 typedef vec4f (das_interop_function) ( das_context * ctx, das_node * node, vec4f * arguments );
